Review Assignment Completed
Forgot to add in detail what I changed. But basically I just added
concatenation, prompts, and validation for everything that didn’t
already have it. Also I fixed my first few flow charts.
Im aware there is a lot more that could be done, but I feel like for
the older assignments it would defeat the purpose of going back and
adding in everything that we learned because I would end up redoing the
whole assignment.